gamification/requirements.txt
asgiref==3.5.2
certifi==2022.9.14
click==8.1.3
Django==4.1.1
django-cors-headers==3.7.0
django-elasticsearch-dsl==7.2.0
django-elasticsearch-dsl-drf==0.22.2
django-environ==0.9.0
django-groupadmin-users==0.3.4
django-mssql-backend==2.8.1
django-nine==0.2.5
django-postgres-extensions==0.9.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
elasticsearch==7.17.6
elasticsearch-dsl==7.4.0
Flask==2.2.2
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
psycopg2-binary==2.9.3
PyJWT==2.4.0
pyodbc==4.0.34
python-dateutil==2.8.2
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.12
Werkzeug==2.2.2
